packages/gooddata-dbt/pyproject.toml
# (C) 2025 GoodData Corporation
[project]
name = "gooddata-dbt"
version = "1.53.0"
description = "dbt plugin for GoodData"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "GoodData", email = "support@gooddata.com"}
]
requires-python = ">=3.9.0"
dependencies = [
    "gooddata-sdk~=1.53.0",
    "pyyaml>=6.0",
    "attrs>=21.4.0,<=24.2.0",
    "cattrs>=22.1.0,<=24.1.1",
    "requests~=2.32.0",
    "tabulate~=0.8.10",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development",
    "Typing :: Typed",
]

[project.scripts]
gooddata-dbt = "gooddata_dbt.main:main"

[dependency-groups]
test = [
    "pytest~=8.3.4",
    "pytest-cov~=6.0.0",
]
type = [
    "mypy~=1.11.2",
    "pydantic~=2.9.2"
]

[tool.hatch.build.targets.wheel]
packages = ["src/gooddata_dbt"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
